requirements.txt
PyMySQL==0.8.1
mysql-replication==0.18
pika==0.12